Add a definition for different types of interesting findings.
The classes that satisfy `mutation_type` describe the different kinds of
difference that can be found.

The results fully defined here are `Snp` for Single Nucleotide
Polymorphisms and the generic change.

All different types can be collected in a `variant` type and searched
with the `check<type>` functions defined here also.